requirements.txt
dash
pandas
matplotlib==3.5.1
gunicorn==20.0.4
dash-bootstrap-components==1.1.0 
dash_bootstrap_templates 
whitenoise==5.2.0
nsetools
requests
pymongo[srv]
flask-restful
